refactor(fuji): extract shared parseDateTime and matchesEntrySearch helpers
Unify the pipe-delimited DateTimeString parser (duplicated in 4 files)
and entry search predicate (duplicated in sidebar + table) into
$lib/utils/ so search semantics stay consistent across views.
